+++ dom/src/lib.rs
@@ -1,3 +1,7 @@
-mod dom;
+pub mod dom;
+pub mod dom_adapter;
 
-pub use dom::*;
+pub mod prelude {
+    pub use crate::dom::*;
+    pub use crate::dom_adapter::*;
+}
