+++ renderer/src/elements/svg.rs
@@ -1,5 +1,5 @@
 use dioxus_native_core::real_dom::NodeImmutable;
-use freya_dom::DioxusNode;
+use freya_dom::prelude::DioxusNode;
 use freya_node_state::Style;
 use skia_safe::{svg, Canvas};
 use torin::geometry::Area;
