+++ renderer/src/window.rs
@@ -2,7 +2,7 @@ use dioxus_native_core::NodeId;
 use freya_common::EventMessage;
 use freya_core::process_render;
 use freya_core::{process_layout, ViewportsCollection};
-use freya_dom::FreyaDOM;
+use freya_dom::prelude::FreyaDOM;
 use freya_layout::Layers;
 use std::ffi::CString;
 use std::num::NonZeroU32;
